backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0

langchain==0.2.6
langgraph==0.2.19
langchain-community==0.2.6
langchain-huggingface==0.1.0

faiss-cpu==1.8.0.post1
sentence-transformers==2.7.0

trafilatura==1.7.0
google-generativeai==0.7.2
python-dotenv==1.0.1
requests==2.32.3